Разделил тесты по 3 файлам добавил в ветку файл confest с фикстурами Так же добавил toml lock и flake
