+++ .flake8
@@ -1,0 +1,3 @@
+[flake8]
+max-line-length = 119
+exclude = .git, __pycache__, venv, .venv
